move PartialPirate to rust
