+++ neofoodclub/odds_change.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING, Any, Iterator
 import dateutil.parser
 from dateutil.tz import UTC
 
+from neofoodclub import PartialPirate
 from neofoodclub.arenas import ARENA_NAMES
-from neofoodclub.pirates import PartialPirate
 
 if TYPE_CHECKING:
     import datetime
